backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
gtts==2.4.0
python-multipart==0.0.6
httpx==0.25.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
